+++ examples/src/test/java/examples/MessageFileExampleTest.java
@@ -6,6 +6,8 @@ import static devopsix.hamcrest.email.MessageMatchers.hasHeader;
 import static devopsix.hamcrest.email.MessageMatchers.hasHeaders;
 import static devopsix.hamcrest.email.MessageMatchers.hasSubject;
 import static devopsix.hamcrest.email.MessageMatchers.hasTo;
+import static devopsix.hamcrest.email.MessageMatchers.hasValidDkimSignature;
+import static java.util.Collections.singletonMap;
 import static org.exparity.hamcrest.date.OffsetDateTimeMatchers.isDay;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.containsString;
@@ -21,6 +23,7 @@ import static org.junit.Assert.assertThat;
 import java.io.IOException;
 import java.io.InputStream;
 import java.time.LocalDate;
+import java.util.Map;
 import java.util.Properties;
 
 import javax.mail.Message;
@@ -45,7 +48,7 @@ public class MessageFileExampleTest {
 
     @Before
     public void loadMessage() throws IOException, MessagingException {
-        try (InputStream messageStream = getClass().getClassLoader().getResourceAsStream("message.eml")) {
+        try (InputStream messageStream = getClass().getClassLoader().getResourceAsStream("message.txt")) {
             Session session = Session.getDefaultInstance(new Properties());
             message = new MimeMessage(session, messageStream);
         }
@@ -59,7 +62,7 @@ public class MessageFileExampleTest {
     
     @Test
     public void messageShouldHaveDate() {
-        assertThat(message, hasDate(isDay(LocalDate.of(2019, 12, 5))));
+        assertThat(message, hasDate(isDay(LocalDate.of(2019, 12, 8))));
     }
     
     @Test
@@ -74,26 +77,35 @@ public class MessageFileExampleTest {
     
     @Test
     public void messageShouldHaveTo() {
-        assertThat(message, hasTo("devopsix@outlook.com"));
-        assertThat(message, hasTo(containsString("@outlook.com")));
+        assertThat(message, hasTo("test-fbll2@mail-tester.com"));
+        assertThat(message, hasTo(containsString("@mail-tester.com")));
     }
     
     @Test
     public void messageShouldHaveReturnPath() {
-        assertThat(message, hasHeader("Return-Path", "devopsix@gmail.com"));
-        assertThat(message, hasHeader("Return-Path", containsString("@gmail.com")));
+        assertThat(message, hasHeader("Return-Path", emptyOrNullString()));
     }
     
     @Test
     @SuppressWarnings({ "unchecked", "rawtypes" })
     public void messageShouldHaveReceivedHeaders() {
-        assertThat(message, hasHeaders("Received", (Matcher)iterableWithSize(4)));
+        assertThat(message, hasHeaders("Received", (Matcher)iterableWithSize(3)));
         assertThat(message, hasHeaders("Received", (Matcher)everyItem(anyOf(startsWith("from "), startsWith("by ")))));
-        assertThat(message, hasHeaders("Received", (Matcher)hasItems(containsString("outlook.com"), containsString("google.com"))));
+        assertThat(message, hasHeaders("Received", (Matcher)hasItems(containsString("mail-tester.com"), containsString("google.com"))));
     }
     
     @Test
-    public void messageShouldHaveDkimSignature() {
+    public void messageShouldHaveValidDkimSignature() {
         assertThat(message, hasHeader("DKIM-Signature", not(emptyOrNullString())));
+        // This is the DKIM public key as published by Google for gmail.com at the time
+        // the test message was recorded.
+        // In your tests, replace with the key that that was actually used for signing your message.
+        final Map<String, String> gmailPublicKey = singletonMap("20161025._domainkey.gmail.com",
+            "k=rsa; p=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAviPGBk4ZB64UfSqWyAicdR7lodhy"
+            + "tae+EYRQVtKDhM+1mXjEqRtP/pDT3sBhazkmA48n2k5NJUyMEoO8nc2r6sUA+/Dom5jRBZp6qDKJOwj"
+            + "J5R/OpHamlRG+YRJQqRtqEgSiJWG7h7efGYWmh4URhFM9k9+rmG/CwCgwx7Et+c8OMlngaLl04/bPmf"
+            + "pjdEyLWyNimk761CX6KymzYiRDNz1MOJOJ7OzFaS4PFbVLn0m5mf0HVNtBpPwWuCNvaFVflUYxEyblb"
+            + "B6h/oWOPGbzoSgtRA47SHV53SwZjIsVpbq4LxUW9IxAEwYzGcSgZ4n5Q8X8TndowsDUzoccPFGhdwIDAQAB");
+        assertThat(message, hasValidDkimSignature(gmailPublicKey));
     }
 }