+++ CHANGES.md
@@ -1,7 +1,11 @@
 # Hamcrest Mail Changes
 
+## Unreleased
+
+* Added matcher for verifying DKIM signatures.
+
 ## Version 0.0.1 (7 December 2019)
 
 Initial release.
 
-* Added message header matchers
+* Added message header matchers.